Add temperature setting. Add correct formating for Llama2 prompts.
